moved user controller to conroller directory
