+++ src/main/controller/User.java
@@ -1,8 +1,10 @@
-package main.model;
+package main.controller;
 
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
+
+import main.model.LibraryModel;
 
 public class User {
     private String username;
