+++ src/tests/UserTest.java
@@ -3,7 +3,8 @@ package tests;
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import main.model.User;
+
+import main.controller.User;
 
 
 class UserTest {
